+++ python/cudf/cudf/tests/series/accessors/test_struct.py
@@ -3,6 +3,7 @@
 import pytest
 
 import cudf
+from cudf.testing import assert_eq
 
 
 def test_struct_iterate_error():
@@ -33,3 +34,25 @@ def test_struct_field_errors(data):
 
     with pytest.raises(IndexError):
         got.struct.field(100)
+
+
+def test_struct_explode():
+    s = cudf.Series([], dtype=cudf.StructDtype({}))
+    expect = cudf.DataFrame({})
+    assert_eq(expect, s.struct.explode())
+
+    s = cudf.Series(
+        [
+            {"a": 1, "b": "x"},
+            {"a": 2, "b": "y"},
+            {"a": 3, "b": "z"},
+            {"a": 4, "b": "a"},
+        ]
+    )
+    expect = cudf.DataFrame({"a": [1, 2, 3, 4], "b": ["x", "y", "z", "a"]})
+    got = s.struct.explode()
+    assert_eq(expect, got)
+
+    # check that a copy was made:
+    got["a"][0] = 5
+    assert_eq(s.struct.explode(), expect)
