+++ python/cudf/cudf/tests/test_struct.py
@@ -1,6 +1,5 @@
 # Copyright (c) 2020-2025, NVIDIA CORPORATION.
 
-import numpy as np
 import pandas as pd
 import pyarrow as pa
 import pytest
@@ -58,140 +57,3 @@ def test_series_construction_with_nulls():
     got = cudf.Series(expect).to_arrow()
 
     assert expect == got
-
-
-@pytest.mark.parametrize(
-    "fields",
-    [
-        {"a": np.dtype(np.int64)},
-        {"a": np.dtype(np.int64), "b": None},
-        {
-            "a": cudf.ListDtype(np.dtype(np.int64)),
-            "b": cudf.Decimal64Dtype(1, 0),
-        },
-        {
-            "a": cudf.ListDtype(cudf.StructDtype({"b": np.dtype(np.int64)})),
-            "b": cudf.ListDtype(cudf.ListDtype(np.dtype(np.int64))),
-        },
-    ],
-)
-def test_serialize_struct_dtype(fields):
-    dtype = cudf.StructDtype(fields)
-    recreated = dtype.__class__.device_deserialize(*dtype.device_serialize())
-    assert recreated == dtype
-
-
-@pytest.mark.parametrize(
-    "series, expected",
-    [
-        (
-            [
-                {"a": "Hello world", "b": []},
-                {"a": "CUDF", "b": [1, 2, 3], "c": 1},
-                {},
-            ],
-            {"a": "Hello world", "b": [], "c": cudf.NA},
-        ),
-        ([{}], {}),
-        (
-            [{"b": True}, {"a": 1, "c": [1, 2, 3], "d": "1", "b": False}],
-            {"a": cudf.NA, "c": cudf.NA, "d": cudf.NA, "b": True},
-        ),
-    ],
-)
-def test_struct_getitem(series, expected):
-    sr = cudf.Series(series)
-    assert sr[0] == expected
-
-
-@pytest.mark.parametrize(
-    "data, item",
-    [
-        (
-            [
-                {"a": "Hello world", "b": []},
-                {"a": "CUDF", "b": [1, 2, 3], "c": cudf.NA},
-                {"a": "abcde", "b": [4, 5, 6], "c": 9},
-            ],
-            {"a": "Hello world", "b": [], "c": cudf.NA},
-        ),
-        (
-            [
-                {"a": "Hello world", "b": []},
-                {"a": "CUDF", "b": [1, 2, 3], "c": cudf.NA},
-                {"a": "abcde", "b": [4, 5, 6], "c": 9},
-            ],
-            {},
-        ),
-        (
-            [
-                {"a": "Hello world", "b": []},
-                {"a": "CUDF", "b": [1, 2, 3], "c": cudf.NA},
-                {"a": "abcde", "b": [4, 5, 6], "c": 9},
-            ],
-            cudf.NA,
-        ),
-        (
-            [
-                {"a": "Hello world", "b": []},
-                {"a": "CUDF", "b": [1, 2, 3], "c": cudf.NA},
-                {"a": "abcde", "b": [4, 5, 6], "c": 9},
-            ],
-            {"a": "Second element", "b": [1, 2], "c": 1000},
-        ),
-    ],
-)
-def test_struct_setitem(data, item):
-    sr = cudf.Series(data)
-    sr[1] = item
-    data[1] = item
-    expected = cudf.Series(data)
-    assert sr.to_arrow() == expected.to_arrow()
-
-
-def test_struct_explode():
-    s = cudf.Series([], dtype=cudf.StructDtype({}))
-    expect = cudf.DataFrame({})
-    assert_eq(expect, s.struct.explode())
-
-    s = cudf.Series(
-        [
-            {"a": 1, "b": "x"},
-            {"a": 2, "b": "y"},
-            {"a": 3, "b": "z"},
-            {"a": 4, "b": "a"},
-        ]
-    )
-    expect = cudf.DataFrame({"a": [1, 2, 3, 4], "b": ["x", "y", "z", "a"]})
-    got = s.struct.explode()
-    assert_eq(expect, got)
-
-    # check that a copy was made:
-    got["a"][0] = 5
-    assert_eq(s.struct.explode(), expect)
-
-
-def test_dataframe_to_struct():
-    df = cudf.DataFrame()
-    expect = cudf.Series(dtype=cudf.StructDtype({}))
-    got = df.to_struct()
-    assert_eq(expect, got)
-
-    df = cudf.DataFrame({"a": [1, 2, 3], "b": ["x", "y", "z"]})
-    expect = cudf.Series(
-        [{"a": 1, "b": "x"}, {"a": 2, "b": "y"}, {"a": 3, "b": "z"}]
-    )
-    got = df.to_struct()
-    assert_eq(expect, got)
-
-    # check that a copy was made:
-    df["a"][0] = 5
-    assert_eq(got, expect)
-
-    # check that a non-string (but convertible to string) named column can be
-    # converted to struct
-    df = cudf.DataFrame([[1, 2], [3, 4]], columns=[(1, "b"), 0])
-    expect = cudf.Series([{"(1, 'b')": 1, "0": 2}, {"(1, 'b')": 3, "0": 4}])
-    with pytest.warns(UserWarning, match="will be casted"):
-        got = df.to_struct()
-    assert_eq(got, expect)
